Minify all SVGs
All the tweaking I had to do was a result of using SVGs that weren't
intended for integration with the web; all sorts of IDs were conflicting.

To fix this, I pulled all the SVGs (expect the animated one) and minified
them on export, telling illustrator to use unique IDs everywhere.

We need to make sure that any future SVGs used on the website, if not
using custom css (animations), are minified.