Use prefers-color-scheme only when checking for dark theme
We had an issue where on some browsers, many elements wouldn't appear.
This was because when we used prefers-color-scheme, we were checking for
both light and dark themes specifically. If a device didn't support it,
then neither would be chosen.

The solution was to only use prefers-color-scheme when checking for a
dark theme, and providing the light theme as default.