Form data passed to Parent, Names for all Fields
Using a callback, each form page will call handleChange() from the parent Dashboard class. The dashboard's state contains the empty application, which handleChange() will update in the future.

Every text field has a name which will be used as the key when matching with the state (so we know what field to update).

- Booleans (checkboxes),
- Dropdowns,
- and gender/race are not yet supported.