Fix UI bugs
There are a number of UI bugs that appeared on mobile due to some absolute
children elements not having relatively positioned parents.

I also stop rendering the background on mobile because it fails to adapt
to the view if the keyboard is popped up.

There was also another issue on the last page where text would wrap on top
of itself. That was due to line-height being set to 0.