Create FAQ section
Uses a JSON file to get its questions.

I had to separate the component from the other parts of the page. Since
the FAQ section can change size vertically, if it were in the same space
as the background image it would resize the background image, and that
doesn't look very good.

I'll be changing the background image so the background between the FAQ
section and the rest of the page is seamless.